+++ datatypes/sources/DtTarjetaDeCredito.cpp
@@ -1,6 +1,6 @@
 #include "../headers/DtTarjetaDeCredito.h"
 
-DtTarjetaDeCredito::DtTarjetaDeCredito(std::string _financiera, int _descuento, float _precio, int _cantAsientos, DtUsuario user) : DtReserva(_precio, _cantAsientos, user) {
+DtTarjetaDeCredito::DtTarjetaDeCredito(std::string _financiera, float _descuento, float _precio, int _cantAsientos, DtUsuario user) : DtReserva(_precio, _cantAsientos, user) {
 	this->Financiera = _financiera;
 	this->PorcentajeDescuento = _descuento;
 }
@@ -9,7 +9,7 @@ std::string DtTarjetaDeCredito::getFinanciera() {
 	return this->Financiera;
 }
 
-int DtTarjetaDeCredito::getDescuento()
+float DtTarjetaDeCredito::getDescuento()
 {
 	return this->PorcentajeDescuento;
 }
