+++ interfaces/ISistema.h
@@ -20,8 +20,8 @@ class ISistema {
 		virtual void AltaComentario(vector<int> padres, string texto, string pelicula, string autor) = 0;
 		virtual void AltaPuntaje(int puntuacion, string pelicula, string usuario) = 0;
         virtual DtUsuario* iniciarSesion(string , string) = 0;
-        virtual void CrearReserva(int cantAsientos, float costo, string titulo, int IdFuncion, string usuario, string banco, string financiera, int descuento) = 0;
-		virtual int ObtenerDescuentoFinanciera(string financiera) = 0;
+        virtual void CrearReserva(int cantAsientos, float costo, string titulo, int IdFuncion, string usuario, string banco, string financiera, float descuento) = 0;
+		virtual double ObtenerDescuentoFinanciera(string financiera) = 0;
         virtual ICollection* ListarTitulos() = 0;
         virtual DtPelicula* SeleccionarPelicula(string) = 0;
 		virtual void VerComentariosyPuntajes(string titulo) = 0;
